Extent config type.

+++ src/draw/camera.cpp
@@ -50,9 +50,9 @@ glm::vec3 Quaternion2EulerAngles(const glm::quat& quaternion)
 }
 #endif
 
-const double PI = 3.14159265358979323846;
-const double rad2degree = 180.0f / PI;
-const double degree2rad = PI / 180.0f;
+const float PI = 3.14159265358979323846f;
+const float rad2degree = 180.0f / PI;
+const float degree2rad = PI / 180.0f;
 void Camera::UpdateViewMatrix(glm::quat& quaternion)
 {
 	if (m_spherical_surface_rotation)
@@ -192,7 +192,7 @@ void Camera::RenderController()
 			UpdatePerspectiveProjectionMatrix();
 		}
 
-		if (ImGui::SliderFloat("Near Plain", &m_near, 0.1, 5.0f))
+		if (ImGui::SliderFloat("Near Plain", &m_near, 0.1f, 5.0f))
 		{
 			if (m_camera_projection == PERSPECTIVE)
 				UpdatePerspectiveProjectionMatrix();
